Refactor: minor changes
