+++ src/App.jsx
@@ -1,8 +1,6 @@
 import { BrowserRouter as Router } from "react-router-dom";
-import { Footer } from "./components";
+import { Navbar, Footer } from "./components";
 import { AppRoutes } from "./routes/Routes";
-import { StepByStepList } from "./components/Steplist/StepList";
-import { Footer } from "./components/Footer/Footer";
 
 function App() {
   return (
